src/requirements.txt
aiobotocore==2.1.1
aiohttp==3.8.1
aioitertools==0.9.0
aiosignal==1.2.0
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
argh==0.26.2
asn1crypto==0.22.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
avro==1.8.2
Babel==2.4.0
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.2.0
billiard==3.3.0.23
boltons==19.1.0
boto==2.48.0
boto3==1.20.21
botocore==1.23.24
celery==3.1.15
certifi==2017.7.27.1
cffi==1.12.3
chardet==3.0.4
charset-normalizer==2.0.10
contextlib2==0.5.5
cramjam==2.5.0
cryptography==3.4.7
defusedxml==0.6.0
Django==1.9
django-allauth==0.32.0
django-appconf==1.0.2
django-bootstrap-form==3.3
django-bootstrap3==7.1.0
django-bower==5.2.0
django-celery==3.2.1
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-environ==0.4.5
django-fields==0.3.0
django-fitbit==0.3.0
django-ios-notifications==0.2.0
django-jet==1.0.6
django-mobile==0.7.0
django-nose==1.4.4
django-redis-cache==1.7.1
django-ses==0.8.2
django-snippetscream==0.0.7
django-sslserver==0.22
django-statsd==2.0.1
django-storages==1.6.5
django-supervisor==0.4.0
django-tastypie==0.14.2
django-test-without-migrations==0.6
djrill==2.1.0
docker==5.0.0
docopt==0.4.0
docutils==0.15.2
ecdsa==0.13
elasticsearch==6.4.0
english==2020.7.0
enum34==1.1.6
Fabric==1.13.2
factory-boy==2.4.0
Faker==0.7.18
fakeredis==0.8.2
fastparquet==0.8.0
feedparser==5.2.1
filechunkio==1.8
firebase-token-generator==2.0.1
fitbit==0.3.0
flower==0.9.2
frozenlist==1.3.0
fsspec==2022.1.0
funcsigs==1.0.2
future==0.18.2
futures==3.1.1
google-api-python-client==1.6.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
JayDeBeApi==1.1.1
jmespath==0.10.0
JPype1==0.6.3
json2html==1.3.0
jsonfield==2.0.2
jsonpickle==1.2
jws==0.1.3
jwt==1.2.0
kombu==3.0.37
leaderboard==3.7.0
longerusername==0.4
mandrill==1.0.57
meld3==1.0.2
mock==2.0.0
monotonic==1.6
mrjob==0.5.10
multidict==6.0.2
mysqlclient==1.4.2.post1
ndg-httpsclient==0.4.2
neotime==1.7.4
newrelic==2.90.0.75
nose==1.3.7
numpy==1.20.3
oauth2client==4.1.2
oauthlib==2.0.2
packaging==20.9
pandas==1.2.4
pansi==2020.7.3
paramiko==2.2.1
pathtools==0.1.2
pbr==3.1.1
prompt-toolkit==3.0.19
py2neo==2021.1.5
pyarrow==7.0.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
pycryptodome
Pygments==2.9.0
PyHive==0.6.4
PyJWT==1.6.4
pymongo==3.11.4
PyMySQL==1.0.2
PyNaCl==1.1.2
pyOpenSSL==17.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.1
python-mimeparse==1.6.0
python-statsd==2.1.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==3.12
redis==2.10.5
requests==2.26.0
requests-oauthlib==0.8.0
rsa==3.4.2
s3fs==2022.1.0
s3transfer==0.5.0
simplejson==3.11.1
six==1.16.0
sqlparse==0.2.3
supervisor==4.0.4
thrift==0.13.0
tornado==4.5.1
typing-extensions==4.1.1
uritemplate==3.0.0
urllib3==1.26.7
watchdog==0.8.3
wcwidth==0.2.5
websocket-client==1.1.0
wrapt==1.13.3
yarl==1.7.2
